Full Moon 5× should be less controversial
We do lose some pixels, but the authors clearly frame everything in the box that we can display. The only places we lose anything are in the two interstitial pixel art displays, and even there I think it works fine. Feel free to revert if you disagree. Test it in 16:9 with Full Screen AR too 😄